com.eg.libraryappserver.bean.book.EsBook

+++ src/main/java/com/eg/libraryappserver/LibraryAppServerApplication.java
@@ -15,7 +15,9 @@ public class LibraryAppServerApplication {
         //为了隐藏配置文件中秘钥等信息，拷贝覆盖
         File src = new ClassPathResource("hide/application.yml").getFile();
         File dest = new ClassPathResource("application.yml").getFile();
-        FileUtils.copyFile(src, dest);
+        if (src.exists()) {
+            FileUtils.copyFile(src, dest);
+        }
 
         SpringApplication.run(LibraryAppServerApplication.class, args);
     }
